+++ rxandroidble/src/main/java/com/polidea/rxandroidble/helpers/LocationServicesOkObservable.java
@@ -1,26 +1,21 @@
 package com.polidea.rxandroidble.helpers;
 
 
-import android.content.BroadcastReceiver;
 import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
-import android.location.LocationManager;
 import android.support.annotation.NonNull;
 
+import bleshadow.javax.inject.Named;
 import com.polidea.rxandroidble.ClientComponent;
 import com.polidea.rxandroidble.DaggerClientComponent;
-import com.polidea.rxandroidble.internal.util.LocationServicesStatus;
 
 import java.util.UUID;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 import bleshadow.javax.inject.Inject;
 
 import rx.Emitter;
 import rx.Observable;
+import rx.Subscription;
 import rx.functions.Action1;
-import rx.functions.Cancellable;
 import rx.internal.operators.OnSubscribeCreate;
 
 /**
@@ -40,34 +35,15 @@ public class LocationServicesOkObservable extends Observable<Boolean> {
     }
 
     @Inject
-    LocationServicesOkObservable(@NonNull final Context context, @NonNull final LocationServicesStatus locationServicesStatus) {
+    LocationServicesOkObservable(
+            @Named(ClientComponent.NamedBooleanObservables.LOCATION_SERVICES_OK) final Observable<Boolean> locationServicesOkObsImpl
+    ) {
         super(new OnSubscribeCreate<>(
                 new Action1<Emitter<Boolean>>() {
                     @Override
                     public void call(final Emitter<Boolean> emitter) {
-                        final boolean locationProviderOk = locationServicesStatus.isLocationProviderOk();
-                        final AtomicBoolean locationProviderOkAtomicBoolean = new AtomicBoolean(locationProviderOk);
-                        emitter.onNext(locationProviderOk);
-
-                        final BroadcastReceiver broadcastReceiver = new BroadcastReceiver() {
-                            @Override
-                            public void onReceive(Context context, Intent intent) {
-                                final boolean newLocationProviderOkValue = locationServicesStatus.isLocationProviderOk();
-                                final boolean valueChanged = locationProviderOkAtomicBoolean
-                                        .compareAndSet(!newLocationProviderOkValue, newLocationProviderOkValue);
-                                if (valueChanged) {
-                                    emitter.onNext(newLocationProviderOkValue);
-                                }
-                            }
-                        };
-
-                        context.registerReceiver(broadcastReceiver, new IntentFilter(LocationManager.PROVIDERS_CHANGED_ACTION));
-                        emitter.setCancellation(new Cancellable() {
-                            @Override
-                            public void cancel() throws Exception {
-                                context.unregisterReceiver(broadcastReceiver);
-                            }
-                        });
+                        Subscription subscription = locationServicesOkObsImpl.subscribe(emitter);
+                        emitter.setSubscription(subscription);
                     }
                 },
                 Emitter.BackpressureMode.LATEST